move gw->gateway, mw->middleware
